+++ app/auth/_components/signup/addInfoForm.tsx
@@ -6,8 +6,8 @@ import GenderSelect from "./genderSelect"
 import { SocialSignupSchema } from "@/constants/signinSchema"
 import { zodResolver } from "@hookform/resolvers/zod"
 import { useForm } from "react-hook-form"
-import useSocialSignup from "@/hook/useSocialSignup"
-import { SocialSignupResponse } from "@/hook/type"
+import { useSocialSignup } from "@/libs/reactQuery/useAuthMutation"
+import { SocialSignupResponse } from "@/libs/reactQuery/type"
 import useCheckNickname from "@/hook/useCheckNicknameExists"
 import SuccessToFailModal from "@/components/commons/successToFailModal"
