+++ libs/reactQuery/type.ts
@@ -2,6 +2,8 @@ import { ContinentType } from "@/types/continent"
 import { UpdatePost } from "@/types/post"
 import { ThemeKeys } from "@/types/theme"
 import { SortConditionType } from "@/types/sortCondition"
+import { UseMutationResult } from "@tanstack/react-query"
+import { UserRequest } from "@/app/auth/_components/signin/type"
 
 /* post mutation type */
 export type useGetPostProps = {
@@ -39,4 +41,25 @@ export type putCommentRequest = {
     commentId: number
     content: string
     postId: number
+}
+
+/* auth */
+export type UserResponse = {
+    accessToken: string
+    refreshToken: string
+}
+export type SignInProps = {
+    email: string
+    password: string
+}
+export type SigninResult = UseMutationResult<UserResponse, Error, UserRequest> & {
+    isOpen: boolean
+    formState: "success" | "fail" | null
+    handleOverlay: (isOpen: boolean, state?: "success" | "fail" | null) => void
+}
+export type SocialSignupResponse = {
+    email: string
+    isFirst: boolean
+    memberId: number
+    token: UserResponse
 }
